+++ src/eletter/classes.py
@@ -37,7 +37,7 @@ class Address(hr.Address):
 
 
 class Group(hr.Group):
-    """ A e-mail address group """
+    """ An e-mail address group """
 
     def __init__(self, display_name: str, addresses: Iterable[SingleAddress]) -> None:
         super().__init__(
@@ -76,6 +76,7 @@ class ContentTyped:
 class MailItem(ABC):
     """ Base class for all ``eletter`` message components """
 
+    #: Optional value for the item's :mailheader:`Content-ID` header
     content_id: Optional[str] = attr.ib(default=None, kw_only=True)
 
     @abstractmethod
@@ -94,6 +95,38 @@ class MailItem(ABC):
         date: Optional[datetime] = None,
         headers: Optional[Mapping[str, Union[str, Iterable[str]]]] = None,
     ) -> EmailMessage:
+        """
+        Convert the `MailItem` into an `~email.message.EmailMessage` with the
+        item's contents as the payload and with the given subject,
+        :mailheader:`From` address, :mailheader:`To` addresses, and optional
+        other headers.
+
+        :param str subject: The e-mail's :mailheader:`Subject` line
+        :param from_:
+            The e-mail's :mailheader:`From` line.  Note that this argument is
+            spelled with an underscore, as "``from``" is a keyword in Python.
+        :type from_: address or iterable of addresses
+        :param to: The e-mail's :mailheader:`To` line
+        :type to: iterable of addresses
+        :param cc: The e-mail's :mailheader:`CC` line
+        :type cc: iterable of addresses
+        :param bcc: The e-mail's :mailheader:`BCC` line
+        :type bcc: iterable of addresses
+        :param reply_to: The e-mail's :mailheader:`Reply-To` line
+        :type reply_to: address or iterable of addresses
+        :param address sender:
+            The e-mail's :mailheader:`Sender` line.  The address must be a
+            string or `Address`, not a `Group`.
+        :param datetime date: The e-mail's :mailheader:`Date` line
+        :param mapping headers:
+            A collection of additional headers to add to the e-mail.  A header
+            value may be either a single string or an iterable of strings to add
+            multiple headers with the same name.  If you wish to set an
+            otherwise-unsupported address header like :mailheader:`Resent-From`
+            to a list of addresses, use the `format_addresses()` function to
+            first convert the addresses to a string.
+        :rtype: email.message.EmailMessage
+        """
         msg = self._compile()
         msg["Subject"] = subject
         msg["From"] = compile_addresses(from_)
@@ -156,8 +189,8 @@ class Attachment(MailItem):
 @attr.s(auto_attribs=True)
 class TextAttachment(Attachment, ContentTyped):
     """
-    A representation of a textual e-mail attachment.  The content type must
-    have a maintype of :mimetype:`text`.
+    A textual e-mail attachment.  ``content_type`` defaults to ``"text/plain"``
+    and must have a maintype of :mimetype:`text`.
     """
 
     DEFAULT_CONTENT_TYPE = "text/plain"
@@ -219,7 +252,10 @@ class TextAttachment(Attachment, ContentTyped):
 
 @attr.s(auto_attribs=True)
 class BytesAttachment(Attachment, ContentTyped):
-    """ A representation of a binary e-mail attachment """
+    """
+    A binary e-mail attachment.  `content_type` defaults to
+    ``"application/octet-stream"``.
+    """
 
     #: The body of the attachment
     content: bytes
@@ -272,7 +308,7 @@ class BytesAttachment(Attachment, ContentTyped):
 
 @attr.s(auto_attribs=True)
 class EmailAttachment(Attachment):
-    """ A representation of a :mimetype:`message/rfc822` e-mail attachment """
+    """ A :mimetype:`message/rfc822` e-mail attachment """
 
     #: The body of the attachment
     content: EmailMessage
@@ -318,6 +354,13 @@ M = TypeVar("M", bound="Multipart")
 
 @attr.s
 class Multipart(MailItem, MutableSequence[MailItem]):
+    """
+    Base class for all multipart classes.  All such classes are mutable
+    sequences of `MailItem`\\s supporting the usual methods (construction from
+    an iterable, subscription, ``append()``, ``pop()``, etc.).
+    """
+
+    #: The `MailItem`\s contained within the instance
     content: List[MailItem] = attr.ib(factory=list, converter=mail_item_list)
 
     @overload
@@ -359,21 +402,27 @@ class Multipart(MailItem, MutableSequence[MailItem]):
         return len(self.content)
 
     def insert(self, index: int, value: MailItem) -> None:
+        """ :meta private: """
         self.content.insert(index, value)
 
     def append(self, value: MailItem) -> None:
+        """ :meta private: """
         self.content.append(value)
 
     def reverse(self) -> None:
+        """ :meta private: """
         self.content.reverse()
 
     def extend(self, values: Iterable[MailItem]) -> None:
+        """ :meta private: """
         self.content.extend(values)
 
     def pop(self, index: int = -1) -> MailItem:
+        """ :meta private: """
         return self.content.pop(index)
 
     def remove(self, value: MailItem) -> None:
+        """ :meta private: """
         self.content.remove(value)
 
     def __iadd__(self: M, other: Iterable[MailItem]) -> M:
@@ -383,6 +432,49 @@ class Multipart(MailItem, MutableSequence[MailItem]):
 
 @attr.s
 class Alternative(Multipart):
+    """
+    A :mimetype:`multipart/alternative` e-mail payload.  E-mails clients will
+    display the resulting payload by choosing whichever part they support best.
+
+    An `Alternative` instance can be created by combining two or more
+    `MailItem`\\s with the ``|`` operator:
+
+    .. code:: python
+
+        text = TextBody("This is displayed on plain text clients.\\n")
+        html = HTMLBody("<p>This is displayed on graphical clients.<p>\\n")
+
+        alternative = text | html
+
+    Likewise, additional `MailItem`\\s can be added to an `Alternative`
+    instance with the ``|=`` operator:
+
+    .. code:: python
+
+        # Same as above:
+        alternative = Alternative()
+        alternative |= TextBody("This is displayed on plain text clients.\\n")
+        alternative |= HTMLBody("<p>This is displayed on graphical clients.<p>\\n")
+
+    When combining two `Alternative` instances with ``|`` or ``|=``, the
+    contents are "flattened":
+
+    .. code:: python
+
+        # Same as above:
+        txtalt = Alternative([
+            TextBody("This is displayed on plain text clients.\\n")
+        ])
+        htmlalt = Alternative([
+            HTMLBody("<p>This is displayed on graphical clients.<p>\\n")
+        ])
+        alternative = txtalt | htmlalt
+        assert alternative.contents == [
+            TextBody("This is displayed on plain text clients.\\n"),
+            HTMLBody("<p>This is displayed on graphical clients.<p>\\n"),
+        ]
+    """
+
     def _compile(self) -> EmailMessage:
         if not self.content:
             raise ValueError("Cannot compose empty Alternative")
@@ -404,6 +496,58 @@ class Alternative(Multipart):
 
 @attr.s
 class Mixed(Multipart):
+    """
+    A :mimetype:`multipart/mixed` e-mail payload.  E-mails clients will
+    display the resulting payload one part after another, with attachments
+    displayed inline if their ``inline`` attribute is set.
+
+    A `Mixed` instance can be created by combining two or more `MailItem`\\s
+    with the ``&`` operator:
+
+    .. code:: python
+
+        text = TextBody("Look at the pretty kitty!\\n")
+        image = BytesAttachment.from_file("snuffles.jpeg", inline=True)
+        sig = TextBody("Sincerely, Me\\n")
+
+        mixed = text & image & sig
+
+    Likewise, additional `MailItem`\\s can be added to a `Mixed` instance with
+    the ``&=`` operator:
+
+    .. code:: python
+
+        # Same as above:
+        mixed = Mixed()
+        mixed &= TextBody("Look at the pretty kitty!\\n")
+        mixed &= BytesAttachment.from_file("snuffles.jpeg", inline=True)
+        mixed &= TextBody("Sincerely, Me\\n")
+
+    When combining two `Mixed` instances with ``&`` or ``&=``, the contents are
+    "flattened":
+
+    .. code:: python
+
+        part1 = Mixed()
+        part1 &= TextBody("Look at the pretty kitty!\\n")
+        part1 &= BytesAttachment.from_file("snuffles.jpeg", inline=True)
+
+        part2 = Mixed()
+        part2 &= TextBody("Now look at this dog.\\n")
+        part2 &= BytesAttachment.from_file("rags.jpeg", inline=True)
+        part2 &= TextBody("Which one is cuter?\\n")
+
+        mixed = part1 & part2
+
+        assert mixed.contents == [
+            TextBody("Look at the pretty kitty!\\n"),
+            BytesAttachment.from_file("snuffles.jpeg", inline=True),
+            TextBody("Now look at this dog.\\n"),
+            BytesAttachment.from_file("rags.jpeg", inline=True),
+            TextBody("Which one is cuter?\\n"),
+        ]
+    """
+
     def _compile(self) -> EmailMessage:
         if not self.content:
             raise ValueError("Cannot compose empty Mixed")
@@ -425,6 +569,96 @@ class Mixed(Multipart):
 
 @attr.s
 class Related(Multipart):
+    """
+    A :mimetype:`multipart/related` e-mail payload.  E-mail clients will
+    display the part indicated by the `start` parameter, or the first part if
+    `start` is not set.  This part may refer to other parts (e.g., images or
+    CSS stylesheets) by their :mailheader:`Content-ID` headers, which can be
+    generated using `email.utils.make_msgid()`.
+
+    .. note::
+
+        :mailheader:`Content-ID` headers begin & end with angle brackets
+        (``<...>``), which need to be stripped off before including the ID in
+        the starting part.
+
+    A `Related` instance can be created by combining two or more `MailItem`\\s
+    with the ``^`` operator:
+
+    .. code:: python
+
+        from email.utils import make_msgid
+
+        img_cid = make_msgid()
+
+        html = HTMLBody(
+            "<p>Look at the pretty kitty!</p>"
+            f'<img src="{img_cid[1:-1]}"/>"
+            "<p>Isn't he <em>darling</em>?</p>"
+        )
+
+        image = BytesAttachment.from_file("snuffles.jpeg", content_id=img_cid)
+
+        related = html ^ image
+
+    Likewise, additional `MailItem`\\s can be added to a `Related` instance
+    with the ``^=`` operator:
+
+    .. code:: python
+
+        # Same as above:
+
+        img_cid = make_msgid()
+
+        related = Related()
+
+        related ^= HTMLBody(
+            "<p>Look at the pretty kitty!</p>"
+            f'<img src="{img_cid[1:-1]}"/>"
+            "<p>Isn't he <em>darling</em>?</p>"
+        )
+
+        related ^= BytesAttachment.from_file("snuffles.jpeg", content_id=img_cid)
+
+    When combining two `Related` instances with ``^`` or ``^=``, the contents
+    are "flattened":
+
+    .. code:: python
+
+        # Same as above:
+
+        img_cid = make_msgid()
+
+        htmlrel = Related([
+            HTMLBody(
+                "<p>Look at the pretty kitty!</p>"
+                f'<img src="{img_cid[1:-1]}"/>"
+                "<p>Isn't he <em>darling</em>?</p>"
+            )
+        ])
+
+        imgrel = Related([
+            BytesAttachment.from_file("snuffles.jpeg", content_id=img_cid)
+        ])
+
+        related = htmlrel ^ imgrel
+
+        assert related.contents == [
+            HTMLBody(
+                "<p>Look at the pretty kitty!</p>"
+                f'<img src="{img_cid[1:-1]}"/>"
+                "<p>Isn't he <em>darling</em>?</p>"
+            ),
+            BytesAttachment.from_file("snuffles.jpeg", content_id=img_cid),
+        ]
+
+    .. tip::
+
+        You can remember the fact that `Related` uses ``^`` by association with
+        :mailheader:`Content-ID`\\s, which are enclosed in ``<...>``, which
+        look like a sideways ``^``!
+    """
+
     def _compile(self) -> EmailMessage:
         if not self.content:
             raise ValueError("Cannot compose empty Related")
@@ -446,6 +680,9 @@ class Related(Multipart):
 
 @attr.s(auto_attribs=True)
 class TextBody(MailItem):
+    """ A :mimetype:`text/plain` e-mail body """
+
+    #: The plain text body
     content: str
 
     def _compile(self) -> EmailMessage:
@@ -456,6 +693,9 @@ class TextBody(MailItem):
 
 @attr.s(auto_attribs=True)
 class HTMLBody(MailItem):
+    """ A :mimetype:`text/html` e-mail body """
+
+    #: The HTML source of the body
     content: str
 
     def _compile(self) -> EmailMessage: